fix: Show Zeebe-Play version
Currently, the Zeebe-Play version is always "dev" if it is used as a Docker image. The common way with the manifest file (generated by the JAR plugin) doesn't work because Jib doesn't use the JAR but the class files directly.

Solve this issue by using a special file to store the version on building the artifact.
By using the JAR, it should fix the issue that the Zeebe-Play version is not shown.